requirements.txt
altgraph==0.17
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
async-timeout==3.0.1
auto-py-to-exe==2.7.8
autopep8==1.5.4
azure-common==1.1.25
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
bottle==0.12.18
bottle-websocket==0.2.9
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.3
cryptography==3.1.1
defusedxml==0.6.0
distlib==0.3.1
Django==3.1.1
django-allauth==0.42.0
django-cors-headers==3.5.0
django-extensions==3.0.9
django-rest-auth==0.9.5
django-storages==1.10.1
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
drf-yasg==1.17.1
Eel==0.12.4
filelock==3.0.12
future==0.18.2
gevent==20.6.2
gevent-websocket==0.10.1
greenlet==0.4.16
idna==2.10
inflection==0.5.1
isort==5.5.2
itypes==1.2.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
macholib==1.14
MarkupSafe==1.1.1
mccabe==0.6.1
mysqlclient==2.0.1
oauthlib==3.1.0
packaging==20.4
Pillow==7.2.0
pipenv==2020.8.13
protobuf==3.13.0
pycodestyle==2.6.0
pycparser==2.20
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.8
PyJWT==1.7.1
pylint==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
uritemplate==3.0.1
urllib3==1.25.10
virtualenv==20.0.33
virtualenv-clone==0.5.4
Werkzeug==1.0.1
whichcraft==0.6.1
wrapt==1.12.1
zope.event==4.4
zope.interface==5.1.0
